requirements.txt
numpy
matplotlib
tqdm
pillow
tensorflow-gpu
h5py
pandas
